pyproject.toml
[project]
name = "nonebot-plugin-mcqq-server"
version = "1.1.0"
description = "mcqq服主版，采用本地读取log信息的方法的Minecraft Server消息互通的插件。"
authors = [{ name = "KarisAya", email = "karisaya@foxmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "nonebot2 (>=2.4.2,<3.0.0)",
    "mcrcon (>=0.7.0,<0.8.0)",
    "nonebot-plugin-alconna (>=0.59.0,<0.60.0)",
    "nonebot-plugin-uninfo (>=0.8.2,<0.9.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
